pyproject.toml
[project]
name = "artist-mgmt-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Rajan66",email = "rajanmaharjan042@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "environs (>=14.1.1,<15.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
